Animals controller done

+++ db/delete_animal.sql
@@ -1,1 +1,1 @@
-DELETE FROM animals WHERE animal_id = $1;
+DELETE FROM animals WHERE animal_id = $1 AND user_id = $2;
